ScreenY/web-app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pillow==10.4.0
opencv-python==4.10.0.84
watchdog==4.0.1
requests
numpy>=1.24
tzdata>=2024.1
paho-mqtt>=1.6.1
